flag변수로 copyFile테스트
